Reworked the toolchain build scripts
Major changes in the way we build our toolchains. We can now
select between LLVM and GNU but I don't know if GNU is fully
functional (the last time I was using it I couldn't get atomics
to work for the kernel). Everything is also a lot more organized
now so future changes should be easier. I still need to change
files under the patch directory to actual patch files so that in
case of a version change we don't have to completely start from
scratch.