Module initialization with mdev
We are now using the mdev daemon to load up the ahci module. For now we
are loading the pci module twice (we need to check for loaded modules in
modprobe). I will need to optimize some things because right now the
mdev probing is taking up about half a minute.